requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
PyYAML==6.0.3
rich==14.2.0
shellingham==1.5.4
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.4
